+++ src/generators/index.ts
@@ -1,2 +1,3 @@
 export * from './json';
 export * from './natural';
+export * from './sections';